refactor: move TextEditor, EmojiPicker and FileUpload
